+++ wxc_sdk/as_api.py
@@ -5894,6 +5894,7 @@ class AsPersonSettingsApiChild(AsApiChild, base=''):
             ('workspaces', 'outgoingPermission/digitPatterns'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'privacy'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'anonymousCallReject'): ('telephony/config/workspaces', '/'),
+            ('workspaces', 'doNotDisturb'): ('telephony/config/workspaces', '/'),
             ('people', 'agent'): ('telephony/config/people', '/'),
             ('people', 'callBridge'): ('telephony/config/people', '/features/'),
             ('people', 'outgoingPermission/'): ('telephony/config/people', '/'),
@@ -7322,51 +7323,49 @@ class AsCallingBehaviorApi(AsPersonSettingsApiChild):
 
 class AsDndApi(AsPersonSettingsApiChild):
     """
-    API for person's DND settings
+    API for person's DND settings. Also used for workspaces
     """
 
     feature = 'doNotDisturb'
 
-    async def read(self, person_id: str, org_id: str = None) -> DND:
+    async def read(self, entity_id: str, org_id: str = None) -> DND:
         """
-        Read Do Not Disturb Settings for a Person
-        Retrieve a Person's Do Not Disturb Settings
+        Read Do Not Disturb Settings for an entity
+        Retrieve an entity's Do Not Disturb Settings
 
         When enabled, this feature will give all incoming calls the busy treatment. Optionally, you can enable a Ring
         Reminder to play a brief tone on your desktop phone when you receive incoming calls.
 
-        This API requires a full, user, or read-only administrator auth token with a scope of spark-admin:people_read
-        or a user auth token with spark:people_read scope can be used by a person to read their settings.
-        :param person_id: Unique identifier for the person.
-        :type person_id: str
-        :param org_id: Person is in this organization. Only admin users of another organization (such as partners) may
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param org_id: Entity is in this organization. Only admin users of another organization (such as partners) may
         use this parameter as the default is the same organization as the token used to access API.
         :type org_id: str
         :return:
         """
-        ep = self.f_ep(person_id=person_id)
+        ep = self.f_ep(entity_id)
         params = org_id and {'orgId': org_id} or None
         return DND.model_validate(await self.get(ep, params=params))
 
-    async def configure(self, person_id: str, dnd_settings: DND, org_id: str = None):
+    async def configure(self, entity_id: str, dnd_settings: DND, org_id: str = None):
         """
-        Configure Do Not Disturb Settings for a Person
-        Configure a Person's Do Not Disturb Settings
+        Configure Do Not Disturb Settings for an entity
+        Configure an entity's Do Not Disturb Settings
 
         When enabled, this feature will give all incoming calls the busy treatment. Optionally, you can enable a Ring
         Reminder to play a brief tone on your desktop phone when you receive incoming calls.
 
         This API requires a full or user administrator auth token with the spark-admin:people_write scope or a user
-        auth token with spark:people_write scope can be used by a person to update their settings.
+        auth token with spark:people_write scope can be used by an entity to update their settings.
 
-        :param person_id: Unique identifier for the person.
-        :type person_id: str
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
         :param dnd_settings: new setting to be applied
         :type dnd_settings: DND
-        :param org_id: Person is in this organization. Only admin users of another organization (such as partners)
+        :param org_id: Entity is in this organization. Only admin users of another organization (such as partners)
             may use this parameter as the default is the same organization as the token used to access API.
         """
-        ep = self.f_ep(person_id=person_id)
+        ep = self.f_ep(entity_id)
         params = org_id and {'orgId': org_id} or None
         await self.put(ep, params=params, data=dnd_settings.model_dump_json())
 
@@ -20275,6 +20274,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
     call_intercept: AsCallInterceptApi
     call_waiting: AsCallWaitingApi
     caller_id: AsCallerIdApi
+    dnd: AsDndApi
     devices: AsWorkspaceDevicesApi
     forwarding: AsPersonForwardingApi
     monitoring: AsMonitoringApi
@@ -20294,6 +20294,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
         self.call_waiting = AsCallWaitingApi(session=session, selector=ApiSelector.workspace)
         self.caller_id = AsCallerIdApi(session=session, selector=ApiSelector.workspace)
         self.devices = AsWorkspaceDevicesApi(session=session)
+        self.dnd = AsDndApi(session=session, selector=ApiSelector.workspace)
         self.forwarding = AsPersonForwardingApi(session=session, selector=ApiSelector.workspace)
         self.monitoring = AsMonitoringApi(session=session, selector=ApiSelector.workspace)
         self.music_on_hold = AsMusicOnHoldApi(session=session, selector=ApiSelector.workspace)