+++ wxc_sdk/person_settings/common.py
@@ -78,6 +78,7 @@ class PersonSettingsApiChild(ApiChild, base=''):
             ('workspaces', 'outgoingPermission/digitPatterns'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'privacy'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'anonymousCallReject'): ('telephony/config/workspaces', '/'),
+            ('workspaces', 'doNotDisturb'): ('telephony/config/workspaces', '/'),
             ('people', 'agent'): ('telephony/config/people', '/'),
             ('people', 'callBridge'): ('telephony/config/people', '/features/'),
             ('people', 'outgoingPermission/'): ('telephony/config/people', '/'),