+++ wxc_sdk/workspace_settings/__init__.py
@@ -15,6 +15,7 @@ from ..person_settings.call_intercept import CallInterceptApi
 from ..person_settings.call_waiting import CallWaitingApi
 from ..person_settings.caller_id import CallerIdApi
 from ..person_settings.common import ApiSelector
+from ..person_settings.dnd import DndApi
 from ..person_settings.forwarding import PersonForwardingApi
 from ..person_settings.monitoring import MonitoringApi
 from ..person_settings.moh import MusicOnHoldApi
@@ -42,6 +43,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
     call_intercept: CallInterceptApi
     call_waiting: CallWaitingApi
     caller_id: CallerIdApi
+    dnd: DndApi
     devices: WorkspaceDevicesApi
     forwarding: PersonForwardingApi
     monitoring: MonitoringApi
@@ -61,6 +63,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
         self.call_waiting = CallWaitingApi(session=session, selector=ApiSelector.workspace)
         self.caller_id = CallerIdApi(session=session, selector=ApiSelector.workspace)
         self.devices = WorkspaceDevicesApi(session=session)
+        self.dnd = DndApi(session=session, selector=ApiSelector.workspace)
         self.forwarding = PersonForwardingApi(session=session, selector=ApiSelector.workspace)
         self.monitoring = MonitoringApi(session=session, selector=ApiSelector.workspace)
         self.music_on_hold = MusicOnHoldApi(session=session, selector=ApiSelector.workspace)